+++ SearchView.swift
@@ -1,0 +1,22 @@
+//
+//  SearchView.swift
+//  MemoAppSwift
+//
+//  Created by 서지완 on 2023/08/29.
+//
+
+import SwiftUI
+
+struct SearchView: View {
+    var body: some View {
+        NavigationView{
+            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        }
+    }
+}
+
+struct SearchView_Previews: PreviewProvider {
+    static var previews: some View {
+        SearchView()
+    }
+}
